Fix warnings

+++ main/modes/aabb_utils.c
@@ -38,8 +38,8 @@ void drawBox(display_t * disp, box_t box, rgba_pixel_t color, uint32_t scalingFa
  */
 bool boxesCollide(box_t box0, box_t box1, uint32_t scalingFactor)
 {
-    return (box0.x0 < box1.x1 + scalingFactor &&
-            box0.x1  + scalingFactor> box1.x0 &&
-            box0.y0 < box1.y1 + scalingFactor &&
-            box0.y1 + scalingFactor > box1.y0);
+    return (box0.x0 < (int32_t)(box1.x1 + scalingFactor) &&
+            (int32_t)(box0.x1 + scalingFactor) > box1.x0 &&
+            box0.y0 < (int32_t)(box1.y1 + scalingFactor) &&
+            (int32_t)(box0.y1 + scalingFactor) > box1.y0);
 }
